+++ src/main/java/ru/levelup/vetclinic/menu/action/ActionVets/VetRemoveMenuAction.java
@@ -2,6 +2,7 @@ package ru.levelup.vetclinic.menu.action.ActionVets;
 
 import ru.levelup.vetclinic.config.HibernateConfiguration;
 import ru.levelup.vetclinic.menu.Action;
+import ru.levelup.vetclinic.menu.MenuAnimals.ConsoleMenuAnimals;
 import ru.levelup.vetclinic.menu.MenuCustomers.ConsoleMenuCustomers;
 import ru.levelup.vetclinic.menu.MenuVets.ConsoleMenuVets;
 import ru.levelup.vetclinic.repository.VetRepository;
@@ -17,8 +18,13 @@ public class VetRemoveMenuAction implements Action {
 
     @Override
     public void execute() {
+        Integer password = ConsoleMenuAnimals.readInt("Введите пароль");
+        if (password == 23472) {
         String vetPersonnelNumber = ConsoleMenuVets.readString("Введите персональный номер ветеринара");
         vetRepository.remove(vetPersonnelNumber);
         System.out.println("Ветеринар с персональным номером: " + vetPersonnelNumber + " удален!");
+        } else {
+            System.out.println("Пароль не верный");
+        }
     }
 }
